Rename tangerine-form-session reducer file to tangerine-form-session-reducer.ts
